Replace the internal error handling logic by the status code pages stack
