+++ samples/Mvc.Server/ViewModels/Shared/ErrorViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Mvc.Server.ViewModels.Authorization {
+namespace Mvc.Server.ViewModels.Shared {
     public class ErrorViewModel {
         [Display(Name = "Error")]
         public string Error { get; set; }
