--- samples/Mvc.Server/Views/Authorization/Error.cshtml
@@ -1,12 +1,0 @@
-﻿@model ErrorViewModel
-    
-<div class="jumbotron">
-    <h2>Ooooops, something went really bad with your OpenID Connect request! :(</h2>
-    <p class="lead text-left">
-        <strong>@Model.Error</strong>
-
-        @if (!string.IsNullOrEmpty(Model.ErrorDescription)) {
-            <small>@Model.ErrorDescription</small>
-        }
-    </p>
-</div>
