+++ samples/Mvc.Server/Views/_ViewImports.cshtml
@@ -3,5 +3,6 @@
 @using Mvc.Server.ViewModels.Account
 @using Mvc.Server.ViewModels.Authorization
 @using Mvc.Server.ViewModels.Manage
+@using Mvc.Server.ViewModels.Shared
 @using Microsoft.AspNetCore.Identity
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
